+++ scripts/winesapos-upgrade.sh
@@ -433,7 +433,7 @@ echo "Switching Steam back to the 'stable' update channel complete."
 echo "Running 3.2.0 to 3.2.1 upgrades complete."
 
 echo "Running 3.2.1 to 3.3.0 upgrades..."
-kdialog_dbus=$(sudo -E -u ${WINESAPOS_USER_NAME} kdialog --title "winesapOS Upgrade" --progressbar "Running 3.2.1 to 3.3.0 upgrades..." 12 | cut -d" " -f1)
+kdialog_dbus=$(sudo -E -u ${WINESAPOS_USER_NAME} kdialog --title "winesapOS Upgrade" --progressbar "Running 3.2.1 to 3.3.0 upgrades..." 13 | cut -d" " -f1)
 echo "Setting up default text editor..."
 grep -q "EDITOR=nano" /etc/environment
 if [ $? -eq 0 ]; then
@@ -581,6 +581,16 @@ if [ $? -ne 0 ]; then
     echo "Adding improved support for FAT file systems..."
     ${CMD_PACMAN_INSTALL} mtools
     echo "Adding improved support for FAT file systems done."
+fi
+sudo -E -u ${WINESAPOS_USER_NAME} ${qdbus_cmd} ${kdialog_dbus} /ProgressDialog Set org.kde.kdialog.ProgressDialog value 12
+
+pacman -Q | grep reiserfsprogs
+if [ $? -ne 0 ]; then
+    echo "Adding support for the ReiserFS file system..."
+    # 'cmake' is required to build 'reiserfs-defrag' but is not installed with 'base-devel'.
+    ${CMD_PACMAN_INSTALL} reiserfsprogs cmake
+    ${CMD_YAY_INSTALL} reiserfs-defrag
+    echo "Adding support for the ReiserFS file system done."
 fi
 sudo -E -u ${WINESAPOS_USER_NAME} ${qdbus_cmd} ${kdialog_dbus} /ProgressDialog org.kde.kdialog.ProgressDialog.close
 echo "Running 3.2.1 to 3.3.0 upgrades complete."