+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Add support for the F2FS file system.
 - Add support for the SSDFS file system.
 - Add improved support for FAT file systems.
+- Add support for the ReiserFS file system.
 
 ## 3.2.1 - The SteamOS 3.4 Update